[Bronze IV] Title: 이칙연산, Time: 104 ms, Memory: 14120 KB -BaekjoonHub

+++ 백준/Bronze/15726. 이칙연산/이칙연산.java
@@ -7,9 +7,9 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        double a = Double.parseDouble(st.nextToken());
-        double b = Double.parseDouble(st.nextToken());
-        double c = Double.parseDouble(st.nextToken());
-        System.out.println((int) Math.max(a * b / c, a / b * c));
+        int a = Integer.parseInt(st.nextToken());
+        int b = Integer.parseInt(st.nextToken());
+        int c = Integer.parseInt(st.nextToken());
+        System.out.println((int) Math.max((double) a * b / c, (double) a / b * c));
     }
 }
